Move is_registered_for_course_expression to queries.py
Extract the queryset-level registration expression from student_management
utils into a dedicated queries module and update its importers.
